+++ actix-utils/src/lib.rs
@@ -1,4 +1,5 @@
 //! Actix utils - various helper services
+pub mod boxed;
 mod cell;
 pub mod cloneable;
 pub mod counter;